Refactor DesignSystemLink to be named GatsbyLink
