+++ src/pages/assets/web-assets.js
@@ -6,9 +6,9 @@ import BulletedList from "../../components/BulletedList";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
-import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
+import Link from "../../components/GatsbyLink";
 import Heading from "../../components/Heading";
 import Image from "../../components/Image";
 import ListItem from "../../components/ListItem";
