+++ src/pages/components/content.js
@@ -11,10 +11,10 @@ import Code from "../../components/Code";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
-import Link from "../../components/DesignSystemLink";
 import Details from "../../components/Details";
 import Example from "../../components/Example";
 import Fact from "../../components/Fact";
+import Link from "../../components/GatsbyLink";
 import Heading from "../../components/Heading";
 import Image from "../../components/Image";
 import Lede from "../../components/Lede";
