+++ src/pages/components/structure.js
@@ -7,11 +7,11 @@ import Code from "../../components/Code";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
-import Link from "../../components/DesignSystemLink";
 import Example from "../../components/Example";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
 import Footer from "../../components/Footer";
+import Link from "../../components/GatsbyLink";
 import Header from "../../components/Header";
 import Heading from "../../components/Heading";
 import ListItem from "../../components/ListItem";
