+++ src/pages/getting-started.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Breadcrumbs from "../components/Breadcrumbs";
-import Link from "../components/DesignSystemLink";
+import Link from "../components/GatsbyLink";
 import Lede from "../components/Lede";
 import Main from "../components/Main";
 import Page from "../components/Page";
