+++ src/pages/patterns/interacting-with-young-people.js
@@ -7,9 +7,9 @@ import Cite from "../../components/Cite";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
-import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
+import Link from "../../components/GatsbyLink";
 import Heading from "../../components/Heading";
 import Image from "../../components/Image";
 import Lede from "../../components/Lede";
