+++ src/pages/standards/colour.js
@@ -8,9 +8,9 @@ import ColourGreyFigure from "../../components/ColourGreyFigure";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
-import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
+import Link from "../../components/GatsbyLink";
 import Heading from "../../components/Heading";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
