+++ src/pages/standards/spacing-and-layout.js
@@ -5,9 +5,9 @@ import BulletedList from "../../components/BulletedList";
 import Cite from "../../components/Cite";
 import Demo from "../../components/Demo";
 import DemoItem from "../../components/DemoItem";
-import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
+import Link from "../../components/GatsbyLink";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
