+++ src/pages/standards/typography.js
@@ -7,8 +7,8 @@ import Code from "../../components/Code";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
-import Link from "../../components/DesignSystemLink";
 import Figures from "../../components/Figures";
+import Link from "../../components/GatsbyLink";
 import Heading from "../../components/Heading";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
